+++ src/main/java/com/imsweb/layout/LayoutFactory.java
@@ -46,6 +46,11 @@ import com.imsweb.layout.record.fixed.naaccr.NaaccrLayout;
 public final class LayoutFactory {
 
     //XML Layouts - constants for the internal layout IDs
+    public static final String LAYOUT_ID_NAACCR_XML_24 = "naaccr-xml-24"; // shortcut for full abstract
+    public static final String LAYOUT_ID_NAACCR_XML_24_ABSTRACT = "naaccr-xml-24-abstract";
+    public static final String LAYOUT_ID_NAACCR_XML_24_MODIFIED = "naaccr-xml-24-modified";
+    public static final String LAYOUT_ID_NAACCR_XML_24_CONFIDENTIAL = "naaccr-xml-24-confidential";
+    public static final String LAYOUT_ID_NAACCR_XML_24_INCIDENCE = "naaccr-xml-24-incidence";
     public static final String LAYOUT_ID_NAACCR_XML_23 = "naaccr-xml-23"; // shortcut for full abstract
     public static final String LAYOUT_ID_NAACCR_XML_23_ABSTRACT = "naaccr-xml-23-abstract";
     public static final String LAYOUT_ID_NAACCR_XML_23_MODIFIED = "naaccr-xml-23-modified";
@@ -125,6 +130,7 @@ public final class LayoutFactory {
 
     // make sure to put all the "aliases" in this list...
     static {
+        _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_24, LAYOUT_ID_NAACCR_XML_24_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_23, LAYOUT_ID_NAACCR_XML_23_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_22, LAYOUT_ID_NAACCR_XML_22_ABSTRACT);
         _INTERNAL_LAYOUT_ID_ALIASES.put(LAYOUT_ID_NAACCR_XML_21, LAYOUT_ID_NAACCR_XML_21_ABSTRACT);
@@ -146,6 +152,11 @@ public final class LayoutFactory {
     // make sure to add the most recent layouts first, they will be "tried" in that order (important for discovery mechanism)
     static {
         //NAACCR XML
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_24, "NAACCR XML 24 Abstract");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_24_ABSTRACT, "NAACCR XML 24 Abstract");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_24_MODIFIED, "NAACCR XML 24 Modified");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_24_CONFIDENTIAL, "NAACCR XML 24 Confidential");
+        _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_24_INCIDENCE, "NAACCR XML 23 Incidence");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_23, "NAACCR XML 23 Abstract");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_23_ABSTRACT, "NAACCR XML 23 Abstract");
         _INTERNAL_LAYOUTS.put(LAYOUT_ID_NAACCR_XML_23_MODIFIED, "NAACCR XML 23 Modified");